VSPC-171 review comments
