+++ vspace/src/main/java/edu/asu/diging/vspace/web/staff/StaffSearchSpaceController.java
@@ -12,8 +12,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 
 import edu.asu.diging.vspace.core.model.ISpace;
+import edu.asu.diging.vspace.core.services.ISpaceManager;
 import edu.asu.diging.vspace.core.services.IStaffSearchManager;
 import edu.asu.diging.vspace.core.services.impl.model.StaffSearchSpaceResults;
 
@@ -22,6 +26,9 @@ public class StaffSearchSpaceController {
 
     @Autowired
     private IStaffSearchManager staffSearchManager;
+    
+    @Autowired
+    private ISpaceManager spaceManager;
 
     @RequestMapping(value = "/staff/search/space")
     public ResponseEntity<StaffSearchSpaceResults> searchInVspace(
@@ -32,6 +39,26 @@ public class StaffSearchSpaceController {
         StaffSearchSpaceResults staffSearch = new StaffSearchSpaceResults();
         staffSearch.setSpaces(spaceList);
         return new ResponseEntity<StaffSearchSpaceResults>(staffSearch, HttpStatus.OK);
+    }
+    
+    @RequestMapping("/staff/spaces/search")
+    public ResponseEntity<String> search(@RequestParam(value = "term", required = false) String search){
+        List<ISpace> spaces = null;
+        if (search != null && !search.trim().isEmpty()) {
+            spaces = spaceManager.findByName(search);
+        } else {
+            spaces = spaceManager.getAllSpaces();
+        }
+        
+        ObjectMapper mapper = new ObjectMapper();
+        ArrayNode idArray = mapper.createArrayNode();
+        for (ISpace space : spaces) {
+            ObjectNode spaceNode = mapper.createObjectNode();
+            spaceNode.put("id", space.getId());
+            spaceNode.put("name", space.getName());
+            idArray.add(spaceNode);
+        }
+        return new ResponseEntity<String>(idArray.toString(), HttpStatus.OK);
     }
 
     /**
